Fix NEI optional again
